Reworking how facilities_api handles PPMS pagination (#8500)
* Reworking how facilities_api handles PPMS pagination

PPMS gave us some new endpoints that support pagination
this pr removes all of the pagination hacks we have used in the past.

* Fixing the `StatsD` listener to work with the new and old style

* Updating how we handle PPMS to better support pagination

We have had to hack pagination into PPMSs responses. This hasn't worked
well in the past. PPMS has given us a paginated endpoint that we are now
using. This PR removes most of the Hacks and workarounds for pagination,
it also changes the formatter for the PPMS response.

* Changes from PR feedback
* Fixing some failing specs